pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opspilot"
version = "0.1.0"
description = "Production-style AI copilot for engineering and operations workflows."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.34.0,<1.0.0",
    "pydantic-settings>=2.8.0,<3.0.0",
    "langchain>=0.3.0,<0.4.0",
    "langgraph>=0.3.0,<0.4.0",
    "psycopg[binary]>=3.2.0,<4.0.0",
    "pgvector>=0.3.0,<0.4.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.0,<1.0.0",
    "pytest>=8.3.0,<9.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
